- Move types closer to where it is defined and used. - Move tests for `lib/utils` out of the `package_root/tests` folder, closer to where the parent modules are.
