+++ packages/calypso-e2e/src/lib/blocks/block-flows/business-hours.ts
@@ -1,4 +1,4 @@
-import { BlockFlow, EditorContext, PublishedPostContext } from '../../../types';
+import { BlockFlow, EditorContext, PublishedPostContext } from './types';
 
 type DaysOfWeek = 'Mon' | 'Tue' | 'Wed' | 'Thu' | 'Fri' | 'Sat' | 'Sun';
