+++ packages/calypso-e2e/src/lib/blocks/block-flows/contact-form.ts
@@ -1,4 +1,4 @@
-import { BlockFlow, EditorContext, PublishedPostContext } from '../../../types';
+import { BlockFlow, EditorContext, PublishedPostContext } from './types';
 
 interface ConfigurationData {
 	nameLabel: string;
