+++ packages/calypso-e2e/src/lib/blocks/block-flows/post-carousel.ts
@@ -1,4 +1,4 @@
-import { BlockFlow } from '../../../types';
+import { BlockFlow } from './types';
 
 const blockParentSelector = '[aria-label="Block: Post Carousel"]';
