+++ packages/calypso-e2e/src/lib/blocks/block-flows/slideshow.ts
@@ -1,5 +1,5 @@
 import { createTestFile } from '../../../media-helper';
-import { BlockFlow, EditorContext, PublishedPostContext } from '../../../types';
+import { BlockFlow, EditorContext, PublishedPostContext } from './types';
 
 interface ConfigurationData {
 	imagePaths: string[];
