+++ packages/calypso-e2e/src/lib/blocks/block-flows/star-rating.ts
@@ -1,4 +1,4 @@
-import { BlockFlow, EditorContext, PublishedPostContext } from '../../../types';
+import { BlockFlow, EditorContext, PublishedPostContext } from './types';
 
 const wholeRatings = [ 1, 2, 3, 4, 5 ] as const;
 type WholeRating = typeof wholeRatings[ number ];
