+++ packages/calypso-e2e/src/lib/blocks/block-flows/subscribe.ts
@@ -1,4 +1,4 @@
-import { BlockFlow, PublishedPostContext } from '../../../types';
+import { BlockFlow, PublishedPostContext } from './types';
 
 const editorBlockParentSelector = '[aria-label="Block: Subscribe"]';
 const publishedBlockParentSelector = '.wp-block-jetpack-subscriptions';
