+++ packages/calypso-e2e/src/lib/utils/editor-tracks-event-manager.ts
@@ -1,5 +1,5 @@
 import { Locator, Page } from 'playwright';
-import { TracksEvent, TracksEventProperties } from '../../types/editor-tracks.types';
+import { TracksEvent, TracksEventProperties } from './types';
 
 // Exporting just for unit testing
 export const createEventMatchingPredicate = (
