+++ packages/calypso-e2e/src/lib/utils/index.ts
@@ -1,6 +1,10 @@
+/* Utilities */
 export * from './validate-translations';
 export * from './get-test-account-by-feature';
 export * from './translate';
 
 // Other items are exported for unit testing, we only care about the manager class.
 export { EditorTracksEventManager } from './editor-tracks-event-manager';
+
+/* Types */
+export * from './types';
