+++ packages/calypso-e2e/src/lib/utils/test/editor-tracks-event-manager.test.ts
@@ -1,6 +1,6 @@
 import { it, describe, expect } from '@jest/globals';
-import { createEventMatchingPredicate } from '../../../src/lib/utils/editor-tracks-event-manager';
-import { TracksEvent } from '../../../src/types/editor-tracks.types';
+import { createEventMatchingPredicate } from '../editor-tracks-event-manager';
+import { TracksEvent } from '../types';
 
 describe( 'Test: createEventMatchingPredicate', function () {
 	const fakeTracksEvents: TracksEvent[] = [
