--- packages/calypso-e2e/src/types/editor-toolbar-component.types.ts
@@ -1,4 +1,0 @@
-/**
- * Editor preview options.
- */
-export type EditorPreviewOptions = 'Desktop' | 'Mobile' | 'Tablet';
