+++ packages/calypso-e2e/src/types/index.ts
@@ -1,8 +1,4 @@
 export * from './rest-api-client.types';
 export * from './data-helper.types';
-export * from './editor-tracks.types';
 export * from './secrets-manager.types';
-export * from './block-flows.types';
 export * from './media-helper.types';
-export * from './editor-settings-sidebar-component.types';
-export * from './editor-toolbar-component.types';
